requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.2=pypi_0
aiosignal=1.4.0=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.11.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=5.0.1=pypi_0
attrs=25.4.0=pypi_0
blas=2.132=openblas
blas-devel=3.9.0=32_h1ea3ea9_openblas
boto3=1.39.14=pypi_0
botocore=1.39.14=pypi_0
bottleneck=1.5.0=py310hf462985_0
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.5=hb9d3cd8_0
ca-certificates=2025.7.14=hbd8a1cb_0
certifi=2025.7.14=pyhd8ed1ab_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.1=pypi_0
cloudpickle=3.1.1=pypi_0
comm=0.2.3=pyhe01879c_0
contourpy=1.3.2=pypi_0
curl=8.14.1=hcaaf1f2_0
cycler=0.12.1=pypi_0
cython=0.29.37=pypi_0
d4rl=1.1=dev_0
datasets=3.6.0=pypi_0
debugpy=1.8.11=py310h6a678d5_0
decorator=5.2.1=pyhd8ed1ab_0
dill=0.3.8=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
expat=2.7.1=hecca717_0
farama-notifications=0.0.4=pypi_0
fasteners=0.19=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.59.0=pypi_0
freetype=2.13.3=h4a9f257_0
frozenlist=1.8.0=pypi_0
fsspec=2025.3.0=pypi_0
glfw=2.9.0=pypi_0
gym=0.23.1=pypi_0
gym-notices=0.1.0=pypi_0
gymnasium=1.2.0=pypi_0
h11=0.16.0=pypi_0
h5py=3.14.0=pypi_0
hf-xet=1.2.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=1.0.1=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=6.29.5=py310h06a4308_1
ipython=8.37.0=pyh8f84b5b_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.5.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh31011fe_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h8a1dbc1_1
ld_impl_linux-64=2.44=h1423503_1
libblas=3.9.0=32_h59b9bed_openblas
libcblas=3.9.0=32_he106b2a_openblas
libcurl=8.14.1=h31d0fb7_0
libedit=3.1.20250104=pl5321h7949ede_0
libev=4.33=hd590300_2
libexpat=2.7.1=hecca717_0
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_3
libgcc-ng=15.1.0=h69a702a_3
libgfortran=15.1.0=h69a702a_3
libgfortran5=15.1.0=hcea5267_3
libgomp=15.1.0=h767d61c_3
liblapack=3.9.0=32_h7ac8fdf_openblas
liblapacke=3.9.0=32_he2f377e_openblas
liblzma=5.8.1=hb9d3cd8_2
liblzma-devel=5.8.1=hb9d3cd8_2
libnghttp2=1.58.0=h47da74e_1
libopenblas=0.3.30=pthreads_h94d23a6_1
libpng=1.6.43=h2797004_0
libsodium=1.0.20=h4ab18f5_0
libssh2=1.11.1=h251f7ec_0
libstdcxx=15.1.0=h8f9b012_3
libstdcxx-ng=15.1.0=h4852527_3
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.2.13=h4ab18f5_6
lz4-c=1.10.0=h5888daf_1
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.3=pypi_0
matplotlib-base=3.10.1=py310h68603db_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mjrl=1.0.0=dev_0
mpmath=1.3.0=pypi_0
mujoco-py=2.1.2.14=pypi_0
multidict=6.7.0=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.1.0=pypi_0
ncurses=6.5=h2d0b736_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
numexpr=2.11.0=py310hbaae466_0
numpy=2.2.6=pypi_0
nvidia-cublas-cu12=12.6.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.6.80=pypi_0
nvidia-cuda-nvrtc-cu12=12.6.77=pypi_0
nvidia-cuda-runtime-cu12=12.6.77=pypi_0
nvidia-cudnn-cu12=9.5.1.17=pypi_0
nvidia-cufft-cu12=11.3.0.4=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.7.77=pypi_0
nvidia-cusolver-cu12=11.7.1.2=pypi_0
nvidia-cusparse-cu12=12.5.4.2=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-ml-py=12.575.51=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.6.77=pypi_0
omegaconf=2.3.0=pypi_0
openblas=0.3.30=pthreads_h6ec200e_1
opencv-python=4.12.0.88=pypi_0
openssl=3.5.1=h7b32b05_0
packaging=25.0=pyh29332c3_1
pandas=2.3.1=py310h277b779_0
parso=0.8.4=pyhd8ed1ab_1
patch=2.7.6=h7f98852_1002
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.3.0=pypi_0
pip=25.1.1=pyh8b19718_0
platformdirs=4.3.8=pyhe01879c_0
prettytable=3.16.0=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
propcache=0.4.1=pypi_0
protobuf=6.31.1=pypi_0
psutil=5.9.0=py310h5eee18b_1
pthread-stubs=0.4=hb9d3cd8_1002
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=22.0.0=pypi_0
pybullet=3.2.7=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pyecharts=2.0.8=pypi_0
pygame=2.6.1=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
pynvml=12.0.0=pypi_0
pyparsing=3.2.3=pypi_0
pyrallis=0.3.1=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0post0=py310h06a4308_2
python-tzdata=2025.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py310h6a678d5_0
qhull=2020.2=h434a139_5
readline=8.2=h8c095d6_2
requests=2.32.4=pypi_0
rich=13.9.4=pypi_0
s3transfer=0.13.1=pypi_0
scikit-learn=1.7.1=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=py310h06a4308_3
setuptools=73.0.1=pyhd8ed1ab_0
shellingham=1.5.4=pypi_0
simplejson=3.20.1=pypi_0
six=1.17.0=pyhe01879c_1
sniffio=1.3.1=pypi_0
sqlite=3.50.2=hb25bd0a_1
stack_data=0.6.3=pyhd8ed1ab_1
swankit=0.2.4=pypi_0
swanlab=0.6.7=pypi_0
sympy=1.14.0=pypi_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.22.1=pypi_0
torch=2.7.1=pypi_0
torchaudio=2.7.1=pypi_0
torchvision=0.22.1=pypi_0
tornado=6.5.1=py310h5eee18b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
triton=3.3.1=pypi_0
typer-slim=0.20.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.1=pypi_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
wrapt=1.17.2=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=3.6.0=pypi_0
xz=5.8.1=hbcc6ac9_2
xz-gpl-tools=5.8.1=hbcc6ac9_2
xz-tools=5.8.1=hb9d3cd8_2
yarl=1.22.0=pypi_0
zeromq=4.3.5=h3b0a872_7
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.2.13=h4ab18f5_6
zstd=1.5.6=ha6fb4c9_0
